# Speedup BM25 + filter