--- a/test/benchmark_bm25/cmd/query.go
+++ b/test/benchmark_bm25/cmd/query.go
@@ -16,8 +16,6 @@ import (
 	"errors"
 	"fmt"
 	"log"
-	"math"
-	"strconv"
 	"time"
 
 	"github.com/spf13/cobra"
@@ -68,9 +66,7 @@ var queryCmd = &cobra.Command{
 
 		log.Print("start querying")
 
-		nDCG := 0.
-		hitsAt1 := 0
-		hitsAt5 := 0
+		scores := lib.Scores{}
 		propNameWithId := lib.SanitizePropName(ds.Queries.PropertyWithId)
 		for i, query := range q {
 			// query.Query = strings.Replace(query.Query, `"`, `\"`)
@@ -89,36 +85,11 @@ var queryCmd = &cobra.Command{
 			times = append(times, time.Since(before))
 
 			resultIds := result.Data["Get"].(map[string]interface{})["Fiqa"].([]interface{})
-			fmt.Print()
-
-			IDCG := 0.
-			for j := 0; j < len(query.MatchingIds); j++ {
-				IDCG += 1. / math.Log(float64(j+2.))
+			if err := scores.AddResult(query.MatchingIds, resultIds, propNameWithId); err != nil {
+				return err
 			}
-
-			DCG := 0.
-			for rank, resultId := range resultIds {
-				id, err := strconv.Atoi(resultId.(map[string]interface{})[propNameWithId].(string))
-				for _, matchigId := range query.MatchingIds {
-					if err != nil {
-						return err
-					}
-					if id == matchigId {
-						if rank == 0 {
-							hitsAt1 += 1
-						}
-						if rank < 5 {
-							hitsAt5 += 1
-						}
-						DCG += 1 / math.Log(float64(rank+2))
-
-					}
-				}
-			}
-			nDCG += DCG / IDCG
-
 			if i%1000 == 0 && i > 0 {
-				log.Printf("completed %d/%d queries. nDCG score: %v", i, len(q), nDCG/float64(i))
+				log.Printf("completed %d/%d queries. nDCG score: %v", i, len(q), scores.CurrentNDCG())
 			}
 		}
 
@@ -134,7 +105,7 @@ var queryCmd = &cobra.Command{
 		fmt.Printf("\nObjects imported: %d\n", objCount)
 		stat := lib.AnalyzeLatencies(times)
 		stat.PrettyPrint()
-		log.Printf("nDCG score: %v, hits at 1: %v, hits at 5: %v", nDCG/float64(len(q)), hitsAt1, hitsAt5)
+		scores.PrettyPrint()
 
 		return nil
 	},