--- a/test/benchmark_bm25/cmd/root.go
+++ b/test/benchmark_bm25/cmd/root.go
@@ -19,21 +19,21 @@ import (
 )
 
 var (
-	Origin             string
-	DatasetConfigPath  string
-	BatchSize          int
-	QueriesCount       int
-	MultiplyProperties int
-	FilterProperties   bool
+	Origin                 string
+	DatasetConfigPath      string
+	BatchSize              int
+	QueriesCount           int
+	MultiplyProperties     int
+	FilterObjectPercentage int
 )
 
 const (
-	DefaultOrigin             = "http://localhost:8080"
-	DefaultDatasetConfigPath  = "datasets.yml"
-	DefaultBatchSize          = 100
-	DefaultQueriesCount       = -1
-	DefaultMultiplyProperties = 1
-	DefaultFilterProperties   = false
+	DefaultOrigin                 = "http://localhost:8080"
+	DefaultDatasetConfigPath      = "datasets.yml"
+	DefaultBatchSize              = 100
+	DefaultQueriesCount           = -1
+	DefaultMultiplyProperties     = 1
+	DefaultFilterObjectPercentage = 0
 )
 
 var rootCmd = &cobra.Command{