### What's being changed:
A bit smarter method of keeping track how many unique documents there are.
Note has the commits from the BM25 benchmark with filtering PR

Fiqa, first 1000queries with a filter of 10%.
Before:
```
Min:    2.660833ms
Mean:   21.10613ms
Max:   63.336667ms
p50:   19.731458ms
p90:       31.18ms
p99:   43.788458ms
```

After:
```
Min:    2.683334ms
Mean:  19.060891ms
Max:    62.70525ms
p50:   18.223333ms
p90:   27.700417ms
p99:      38.537ms
```